refactor: split _search_service.py into _search_mixin.py and _reconcile_mixin.py
SearchMixin search/query methods extracted to _search_mixin.py (~250 lines).
ReconMixin repair/reconciliation methods extracted to _reconcile_mixin.py (~356 lines).
_search_service.py kept as re-export for backward compatibility.
service.py updated to inherit from ReconMixin.
